final commit

+++ database/migrations/2022_05_13_220658_create_affiliates_table.php
@@ -18,7 +18,8 @@ return new class extends Migration
             $table->foreignId('user_id');
             $table->foreignId('merchant_id');
             // TODO: Replace me with a brief explanation of why floats aren't the correct data type, and replace with the correct data type.
-            $table->float('commission_rate');
+            // floating point numbers have hard time representing certian number with precision thus can result in calculation errors so use decimal instead, moreover assertEquals doesn't work on float type in some case
+            $table->decimal('commission_rate');
             $table->string('discount_code');
             $table->timestamps();
         });
